// src/app/models/event.ts
export interface IEvent {
  id: number;
  name: string;
  date: Date;
  time: string;
  price: number;
  imageUrl: string;
  location?: {
    address: string;
    city: string;
    country: string;
  };
  onlineUrl?: string;
  sessions?: ISession[];
}

export interface ISession {
  id: number;
  name: string;
  presenter: string;
  durarion: number;
  level: string;
  abstract: string;
  voters: string[];
}